requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
async-timeout==4.0.2
attrs==21.4.0
auto-py-to-exe==2.20.1
bottle==0.12.21
bottle-websocket==0.2.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
Eel==0.14.0
frozenlist==1.3.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
multidict==6.0.2
pefile==2022.5.30
playwright==1.23.0
pycparser==2.21
pyee==8.1.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pyparsing==3.0.9
pywin32-ctypes==0.2.0
requests==2.28.1
slackclient==2.9.4
urllib3==1.26.9
websockets==10.1
whichcraft==0.6.1
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
